Stored the logo for the home panel.
Added a home page in the class HomePage that makes the object a JPanel. Added a side menu in the class SidePanelMenu that makes the object a JPanel. A standard GUI class with a extra private method 'credits' that shows the creators of the application. Added a main TODO list in the class UserInterFace for the whole application.